refactor: move PublicationMetadata to types module and update imports
